// src/app/core/api/models/quarterly-update.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { SubmissionDetails } from '../models/submission-details';
export interface QuarterlyUpdate {
  allowableExpenses?: number;
  businessId?: number;
  id?: string | null;
  netProfit?: number;
  quarterName?: string | null;
  status?: string | null;
  submissionDetails?: SubmissionDetails;
  taxYear?: string | null;
  taxableIncome?: number;
}